changed item amount to be a double and fixed add_item to actually add an item to database

+++ lib/database_helpers.dart
@@ -61,7 +61,7 @@ class DatabaseHelper {
               CREATE TABLE IF NOT EXISTS $TABLE_ITEMS (
                 $COLUMN_ID INTEGER PRIMARY KEY,
                 $COLUMN_NAME TEXT NOT NULL,
-                $COLUMN_AMOUNT INTEGER NOT NULL,
+                $COLUMN_AMOUNT FLOAT NOT NULL,
                 $COLUMN_TRANSACTION_TYPE BOOLEAN NOT NULL,
                 $COLUMN_ACCOUNT_ID INTEGER NOT NULL,
                 $COLUMN_CATEGORY_ID INTEGER,
